backend/pyproject.toml
[project]
name = "nutrilens-backend"
version = "0.1.0"
description = "NutriLens — Backend FastAPI com pipeline de visão computacional para diário nutricional."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.12",
    "pillow>=11.0.0",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.10.0",
    "rapidfuzz>=3.10.0",
    # IA / ML (descomente conforme necessário)
    # "ultralytics>=8.3.0",       # YOLOv11
    # "segment-anything-2",       # SAM 2
    # "google-generativeai>=0.8.0", # Gemini
    # "anthropic>=0.39.0",        # Claude
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
